+++ src/Config/SimpleConfig.php
@@ -14,55 +14,16 @@ class SimpleConfig implements \ArrayAccess, \Iterator
 
     /**
      * @param array $config
-     *
-     * @api
-     */
-    public function __construct(array $config = [])
-    {
-        $this->config = $config;
-    }
-
-    /**
-     * @deprecated internals
-     * @todo will removed since 2.0 version
-     *
      * @param array $placeholders
      *
-     * @return $this
-     *
      * @api
      */
-    public function replace(array $placeholders)
+    public function __construct(array $config = [], array $placeholders = [])
     {
-        trigger_error(sprintf('%s is deprecated.', __METHOD__), E_USER_DEPRECATED);
-        if (isset($this->config['parameters'])) {
-            ArrayHelper::transform($this->config['parameters'], $placeholders);
-            $placeholders = array_merge($this->config['parameters'], $placeholders);
-            unset($this->config['parameters']);
+        $this->config = $config;
+        if ($config && (!empty($config['parameters']) || !empty($placeholders))) {
+            $this->transform($placeholders);
         }
-        ArrayHelper::transform($this->config, $placeholders);
-        return $this;
-    }
-
-    /**
-     * @deprecated use {@link ArrayAccess} interface
-     * @todo will removed since 2.0 version
-     *
-     * @return array
-     *
-     * @api
-     */
-    public function toArray()
-    {
-        trigger_error(sprintf('%s is deprecated.', __METHOD__), E_USER_DEPRECATED);
-        if (isset($this->config['imports'])) {
-            unset($this->config['imports']);
-        }
-        if (isset($this->config['parameters'])) {
-            ArrayHelper::transform($this->config, $this->config['parameters']);
-            unset($this->config['parameters']);
-        }
-        return $this->config;
     }
 
     /**
@@ -72,7 +33,7 @@ class SimpleConfig implements \ArrayAccess, \Iterator
      */
     public function offsetExists($offset)
     {
-        return (bool) ArrayHelper::findByPath($this->config, $offset);
+        return (bool)ArrayHelper::findByPath($this->config, $offset);
     }
 
     /**
@@ -86,8 +47,7 @@ class SimpleConfig implements \ArrayAccess, \Iterator
     }
 
     /**
-     * @param mixed $offset
-     * @param mixed $value
+     * {@inheritdoc}
      *
      * @throws \BadMethodCallException
      *
@@ -99,7 +59,7 @@ class SimpleConfig implements \ArrayAccess, \Iterator
     }
 
     /**
-     * @param mixed $offset
+     * {@inheritdoc}
      *
      * @throws \BadMethodCallException
      *
@@ -158,5 +118,22 @@ class SimpleConfig implements \ArrayAccess, \Iterator
     public function rewind()
     {
         reset($this->config);
+    }
+
+    /**
+     * @param array $placeholders
+     *
+     * @return $this
+     */
+    protected function transform(array $placeholders)
+    {
+        if (isset($this->config['parameters'])) {
+            ArrayHelper::transform($this->config['parameters'], $placeholders);
+            $placeholders = array_merge($this->config['parameters'], $placeholders);
+            unset($this->config['parameters']);
+        }
+        ArrayHelper::transform($this->config, $placeholders);
+        $this->rewind();
+        return $this;
     }
 }