+++ tests/Doctrine/Util/ConfigResolverTest.php
@@ -4,7 +4,7 @@ namespace OctoLab\Common\Doctrine\Util;
 
 use Doctrine\DBAL\Types\Type;
 use OctoLab\Common\Config\Loader;
-use OctoLab\Common\Config\YamlConfig;
+use OctoLab\Common\Config\FileConfig;
 use OctoLab\Common\TestCase;
 use Symfony\Component\Config\FileLocator;
 
@@ -18,12 +18,11 @@ class ConfigResolverTest extends TestCase
      */
     public function resolve()
     {
-        $config = (new YamlConfig(new Loader\FileLoader(new FileLocator(), new Loader\Parser\YamlParser())))
+        $config = (new FileConfig(new Loader\FileLoader(new FileLocator(), new Loader\Parser\YamlParser())))
             ->load($this->getConfigPath('doctrine/config'))
-            ->toArray()
         ;
         $resolver = new ConfigResolver();
-        $resolver->resolve($config['doctrine']['dbal']);
+        $resolver->resolve($config['doctrine:dbal']);
         self::assertTrue(Type::hasType('enum'));
     }
 }